+++ lib/mode/dialog_report_mode.dart
@@ -1,5 +1,5 @@
-import 'package:catcher/model/report.dart';
 import 'package:catcher/model/report_mode.dart';
+import 'package:catcher/model/report.dart';
 import 'package:flutter/material.dart';
 
 class DialogReportMode extends ReportMode {
@@ -8,9 +8,8 @@ class DialogReportMode extends ReportMode {
     _showDialog(report, context);
   }
 
-  Future<void> _showDialog(Report report, BuildContext context) async {
+  _showDialog(Report report, BuildContext context) async {
     await Future.delayed(Duration.zero);
-
     showDialog(
         context: context,
         builder: (BuildContext build) {