+++ lib/mode/page_report_mode.dart
@@ -12,17 +12,17 @@ class PageReportMode extends ReportMode {
 
   @override
   void requestAction(Report report, BuildContext context) {
-    _navigate(report, context);
+    _navigateToPageWidget(report, context);
   }
 
-  Future<void> _navigate(Report report, BuildContext context) async {
+  _navigateToPageWidget(Report report, BuildContext context) async {
     await Future.delayed(Duration.zero);
-
     Navigator.push(
       context,
       MaterialPageRoute(builder: (context) => PageWidget(this, report)),
     );
   }
+
 }
 
 class PageWidget extends StatefulWidget {